コードを整理
